genericcli: do not send \n if answer ends with null-byte
comocutor answers previously had an explicit attribute

```python
  send_nl: bool
```

which controlled whether or not newline is sent after the answer.
gnetcli does not support it, but it is nessesary for some question-promts.
Namely it is needed for cisco catalyst's no username ... command, with question:

  This operation will remove all username related configurations with same name.Do you want to continue? [confirm] ::: Y

If we respond to it with "Y\n" instead of "Y" this will result in duplicated command prompt returned by a device:

```
  "Y\r\ndevice-01(config)#\r\ndevice-01(config)#"
```

If both of them is read during a single GenericReadX() call everything is fine.

But if the next command is issued while only the first prompt has been read this results the second prompt appearing seemingly as a result of a new command.

This means that any command following "no username ... / This operation / Y" has a chance to end up failing with a EchoReadException since:

 * the command has been issued
 * a valid prompt has been read and matched
 * no echo of the issued command has been read

Because of it we need working send_nl=0 and the easiest way to implement it is to force every client add \n to all QA's.
But since it would require to rework every single gnetcli client easier to add an additional semantic to a null byte.
This way %send_nl=0 can be handled in annet in a following manner:

```
% git diff
diff --git a/annet/deploy.py b/annet/deploy.py
index 9fd2ab8..e88fef4 100644
--- a/annet/deploy.py
+++ b/annet/deploy.py
@@ -157,14 +157,15 @@ class RulebookQuestionHandler:

 def rb_question_to_question(q: MakeMessageMatcher, a: Answer) -> Question:  # TODO: drop MakeMessageMatcher
-    if not a.send_nl:
-        raise Exception("not supported false send_nl")
     text: str = q._text  # pylint: disable=protected-access
+    answer: str = a.text
     is_regexp = False
     if text.startswith("/") and text.endswith("/"):
         is_regexp = True
         text = text[1:-1]
-    res = Question(question=text, answer=a.text, is_regexp=is_regexp)
+    if not a.send_nl:
+        answer += "\x00"
+    res = Question(question=text, answer=answer, is_regexp=is_regexp)
     return res
```